Refactor the way integration modules introduce their types the queue processor.
The function queue_processor_add_type($type) is now available for integration modules
to call during hook_enable() to tell the processor about the types they support. This
replaces the old method in which the integration modules had to know the name of the
variable that queue_processor uses to store the processing order. This is a step in
the right direction.