Add a form validation error message when a donation fails at the payment gateway.
We've added a call to form_set_error() when the donation is declined at the payment
gateway so that donations made via the API can catch this error just like any
other client valiation error.

This means that on the Drupal site it is possible that the message will show up
twice. Once via drupal_set_message and once via form_set_error. It probably makes
the most sense to get rid of the drupal_set_message at this point and let it
display like a standard field error.