DO NOT remove spaces from profile_first_name and profile_last_name.
This change is being made because there have been multiple complaints
recently about spaces being removed from first and last name profile
fields.

The origin of this space removing code is unknown. It was thought that in
early implementations there was some issue with spaces in the name
fields in Salesforce.

Who knows. If you want it back, revert this commit.

Cheers :)