Add configuration for boolean validation.
Salesforce requires either a 1 or 0 to be passed to it's API to
represent true or false values. Since many different types of content
pass through the same validation function, a configuration option
has been added to the queue processor settings that allows an admin
to specific what strings represent a TRUE boolean value.

The default values are 'true', 'yes', 'y' and '1'.