Fix for the 'manage own recurring donations' permission.
Now it actually checks that the logged-in $user is the same as the $account that was loaded from the URL. Before this fix, it was checking that $account had the 'manage own' permission, instead of the logged-in user.

Also added in the 'cancel future donations' permission for administrators/managers to be able to use this page.