Update the page_wrappers table when the page_wrapper node is updated.
Previously, the assumption was made that the insert operation would insert a valid string to the template column, and the update operation would never need to change that value.

However, if the template file creation fails for any reason, _page_wrappers_create_template() returns FALSE. This FALSE value could never be changed.