Bug fix for incorrect stage name being sent to Salesforce when recurring donations are updated.
The stage name does not need to be manually set for recurring donations. The default stage
is already set to pledged in donation.php when an instance of a donation is created. If there
is a payment for the donation, the stage will get set to Posted (NPSP) or Received (Common Groud).