Applying patch from the IFAW fork of webform_user to fix uid 0 profile values bug.
Specifically, this part of IFAW SHA ab9da725e3025b3254ce606cb3307c766cd3c0c2:

"3) Fixes to the email matching logic so that it operates on valid $account objects, and never the anonymous user."

Adding watchdog logging for when you *would have* saved profile values to uid 0, but this code prevented it. That might help troubleshooting in the future.