Add batch size per type.
The batch size is now configurable per type of integration. This
means you can now set one item to be sent to Salesforce in batches
of 200 while others can use a different batch size. This is useful
if you encounter a misbehaving trigger that isn't properly bulkified.
You can dial down the batch size to work around the broken trigger.