Manual merge of pull request 137. https://github.com/JacksonRiver/springboard_modules/pull/137
This pull req is a cumulative update of work done on EL re the extraOptions field for CIM transactions, allowing up to five webform components to be mapped into custom fields in the admin page for a given gateway. This is definitely first draft work, and could use attention later for good UI and user-safe configuration.

In addition, Coder updates and basic cleanup is included on this update.