Changing createBatches() to store the sObject separately from queue items
The calls to Salesforce need the sObject only, so I would be required to
loop through all of the items once more in order to gather up all of the
sObjects.  Instead, I changed createBatches() to store the sObject as
another group separately from the queue items.

I also tweaked the approach to sorting the batches so it would be done
with one loop through the items, and then going through each order of
operations to send to Salesforce and grabbing categorized batches in that
order and placing them in the batch queue.