Adding webform user integration support.
This commit adds an additional submit handler to the
mapping form that will examine the mapped out for a
reference field to the contact object.

It also implements hook_salesforce_genmap_map_fields_alter()
where it adds the foreign key field to the object being
exported using a token as the fields value.