Changing order of processor checks, checking queue items before Salesforce
Because processor is run on cron, and an error is issued if Salesforce
cannot connect, it would be better to first check to see if there are any
items in the queue.  Cron is run on numerous occasions before Salesforce
Queue is ever configured and tons of errors are being sent out in the mean
time.  This way, errors are being sent out only when the queue has started
to be used.