Converting our references to salesforce queues to the controlling queue
This utilizes the patch to the salesforce module that allows addon
modules to invoke task hooks, and the support for pull_queue and push_queue
tasks.  It adds a helper function to make the system still work with our
push queue even if the salesforce module is not yet patched.

Removed the admin interface from salesforce_genmap that allowed you to
choose the queue it used, since this deprecates that.

Move dequeing items related to node delete from salesforce_gen map to
salesforce_queue.

Removed queue crud interface since all queues will assume to use the
DrupalReliableQueueInterface directly with the queue object.