Adding Salesforce NPSP module for features specific to that package.
This module currently implements the post process hook for batches
and will retrieve and store the account ids for Drupal users that
have been exported to Salesforce contacts.

For convenience, it also implements hook_user_load() to add the
user's account and contact ids to the $user object. These are
available as $user->salesforce_contact_id and
$user->salesforce_account_id. It also displays these ids on the
user account view page.