Alter the opportunity object after a sustainer charge to only include fields that were part of the transaction.
The purpose of this change is to prevent other fields on the opportunity
from being updated after a sustainer charge is run. This will allow the
opportunity to be edited in Salesforce and not have it overwritten when
the sustainer charge is processed.

This change sends the basic fields, StageName, Probability and
Transaction_Date_Time__c. They are read out of a variable so that
a config screen can be added later to control while fields get updated.