Remove trailing whitespace
It also helps avoid conflict.  Of course, there is a chance that if you are
reading this, you are saying to yourself, "BUT IT DIDN'T AVOID THIS
CONFLICT!" It does cause problems when trailing white space is altered,
which is why trailing whitespace should always be removed completely
because you can never know when an idle finger swipe on the space bar while
pondering could cause the next guy to come along and pull his hair out
wondering why things are conflicting.

So, I appologize for your current frustration, but hopefully there will be
a lot less of it in the future if we eliminate all trailing whitespace
going forward.

The magic: find . -type f -exec sed -i 's/ *$//' '{}' ';'