Add caching to `fundraiser_commerce_get_countries()`.
This function is somewhat expensive because there is a call to `_fundraiser_commerce_get_zone_by_country()` for every country returned by `_fundraiser_commerce_get_countries()`. Since there are no parameters on the parent function, the return value is the same every time and is cacheable. I added static caching for a single runtime, but also save $return_data in a cache bin.