Adding additional protections to prevent double charging of sustainers.
The sustainer key is now read at donation time and written to
fundraiser_sustainers table. This value is then checked during
cron runs to ensure the key value stored in the sustainer record
matches the configured key.

This will prevent donations that were created in a production
environment from being charged in a dev environment that has
a valid key configured.