Adding UI module to expose some manual sync functionality to users.
This first release allows a permissioned user to delete the sync
map records. It exposes a Salesforce sync tab on the user account
to accomplish this. After the sync map record is deleted the
user can manually put the user back into the queue on the same
tab.

Adding the ability to manually put an item back into the queue. Coder review cleanup.