Escape XML characters in string fields to prevent SOAP faults.
Use htmlspecialchars() function to escape invalid XML characters
such as &. This should suffice for now. If more rigorous escaping
is needed we should explore using DOMDocument or SimpleXML, which
will escape strings properly.

The batchConvertToSalesforceFormat() method was also moved to the
end of the pipeline of events that occur in sendBatchPrepare() to
ensure bad characters aren't added to the object from Salesforce
during the application of the fieldmap directional rules.