Fix fix for salesforce mapping key default value not being set
It mistakenly set all rows as being the key, making the last row actually
be used as the key.  Subsequently, the query builder would mistakenly send
empty values for the key and cause us to try to fetch every single row in
the Salesforce database.  At a million rows, thats about 500k queries.