Fix failing email_wrappers test
After the creation of the test email wrapper, the test wanted to assert
that the email wrapper node was unpublished by seeing that it redirected
the user to the email wrapper node after creation.  However, the test
server redirected to the homepage while everyone else's local environments
redirected to the node view page. This commit changes the method of
checking the publish status by checking against the node directly.