Standardize test documentation in .info files
Originally the intent was to prevent the auto load of setup tests, but
apparently this just isn't possible.  Drupal doesn't support SimpleTest's
ability to hide test cases by setting them in the "abstract" scope.  Then,
for whatever reason, extending classes that are not autoloaded will not
work, even if you explicitely load the file.  Sooo... here we are with
empty tests.

Conflicts:
	fundraiser/modules/fundraiser_commerce/fundraiser_commerce.info
	fundraiser/modules/fundraiser_webform/fundraiser_webform.info

Conflicts:
	fundraiser/modules/fundraiser_webform/fundraiser_webform.info