Sets object record type correctly for items mapped through the Salesforce generic mapping module.
Fixes incorrect record types being set for donations and webform
maps. The genmap module used to store the record type name instead
of it's id. The constructor for SalesforceObject expects it's 3rd
parameter to be a record type name. I've removed the record type
id from the new object instantiation and instead am setting it
explictly using the setRecordTypeId() method. This will correctly
add the RecordTypeId field to the object and set the object's
record type name.