Change CREATE operations to UPDATE when SFID is available.
This update will check queue items that have been inserted
with a CREATE operation to see if they have already been
exported. This check is done by looking for a record in the
sync_map table. If a record with an SFID is found, the
operation is changed from CREATE to UPDATE and the SFID is
appended to the object.

This prevents duplicates from being created in Salesforce.