Prevent SOAP error in batches where an item contains a bad token.
This update fixes the SOAP-ERROR: Encoding: object has no 'type' property
error encountered when a batch contains an item that has a non-replaceable
relationship token. The bad token was causing the entire batch to fail
instead of just the single item because the array of sobjects sent to
Salesforce did not start at zero. This update resets array indices for
both $batch['items'] and $batch['sobjects'] whenever an item is removed
because of a bad token.