Prevents user_save() from being called twice on webform submission.
Alters _webform_user_save_profile_map() to take advantage of the ability of
user_save() to both create new users or update existing ones. Rather than
passing in a UID as a parameter, we pass in an $account object. When creating
a new user, add the available information such as $account->mail, and add
$account->is_new = TRUE.